requirements.txt
# April Agent Dependencies
# Google ADK and core
google-adk>=1.0.0
google-genai>=1.0.0

# MCP support (requires Python 3.10+)
mcp>=1.0.0

# Web server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Database session support (PostgreSQL)
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0

# For Firestore (production user credential storage)
# Uncomment when ready:
# google-cloud-firestore>=2.11.0


